Issue 64: Check buffer length to avoid segfaults
To avoid a Java-level buffer length bug from taking out the entire
JVM with a segfault, perform boundary checks on the length
parameters that are passed on to libsodium.

For most usecases the buffer lengths are probably redundant as we
can just automatically pass .length from Java, removing the need
for the caller to explicitly do so. The existence of the unsigned
length parameters are primarily to identify the end of the buffer
in C; a problem Java does not have.

This change fixes #64